Code/requirements.txt
torch>=2.0.0

numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

biopython>=1.79  

fair-esm>=2.0.0  

tqdm>=4.60.0